Allow empty value
When a split package overriddes an empty global filed makepkg generates
an empty value in the srcinfo. For example `depends = `

I believe this to be a bug but as it is part of makepkg's behavour and
there are countless srcinfos that may suffer from this and it would be
best not to fail while parsing them.