Correctly handle empty overrides.
When a srcinfo contains an empty value such as "pkgdesc=" this means
that the global value should be overriden and set to nothing.

There is a makepkg bug that causes these empty overrides to be
incorrectly generated in the srcinfo. Although it does not happen often
and the fix is already merged. From the perspective of this project,
this bug can be ignored.